### 2018-12-18 #1133 @whedon
[PRE REVIEW]: RSeqAn: Headers and wrappers for the SeqAn library in R
**Submitting author:** @aguang (<a href="http://orcid.org/0000-0003-4974-1873">August Guang</a>)
**Repository:** <a href="https://github.com/compbiocore/RSeqAn" target ="_blank">https://github.com/compbiocore/RSeqAn</a>
**Version:** v1.0
