### 2018-12-29 #1133 @brainstorm
@vladsaveliev, would you like to help JOSS review this software? Shouldn't take long to evaluate, I'll help you out as editor ;)
