### 2019-01-18 #1160 @brainstorm
@aguang, could you please fix the table in the paper so that it displays properly? Thanks @pdiakumis for pointing it out!
