### 2019-01-18 #1160 @pdiakumis
@aguang, I also believe the release version (v1.0 above) needs to match the one in your GitHub repo (v1.3.1 in the [DESCRIPTION](https://github.com/compbiocore/RSeqAn/blob/deae809018536e4f51e4bb496e605586e004756b/DESCRIPTION#L4) file (?)) - thanks!
