### 2019-01-24 #1160 @aguang
@brainstorm @pdiakumis I think I've fixed the table. As for the version, I may have submitted the wrong version number, it should be v1.3.1. Is there a way for me to change that in my submission?
