### 2019-01-30 #1160 @brainstorm
@aguang, could you please add the missing expected fields in the paper.md? Namely `date`/`bibliography` as stated above in the whedon output? Thanks!
