### 2019-02-03 #1160 @pdiakumis
Hi @aguang,

First of all, thank you so much for submitting to JOSS!

It was my very first review, and my Rcpp/C++/SeqAn knowledge was non-existent, so apologies for the delay. On the bright side, I learnt how to use them over the afternoon and I've been blown away by the magic.

Here are a couple of comments:

* Maybe a quick note under [Usage](https://github.com/compbiocore/RSeqAn#usage) with e.g. "for a quick test/example, use `Rcpp::sourceCpp("RSeqAn_example1.cpp")`" would be great for people like myself, who might be experienced with R but not Rcpp/C++. That `Sys.setenv` note was also very helpful.

* The version issue seems to have been addressed.
* I think the only outstanding issue is with the PDF compilation, so for that see the error output above. If that gets solved, then I believe we're good to go @brainstorm (?).

Kind Regards,
Peter
