### 2019-02-05 6c5882c
updated date and bibliography in yaml and added direct source code in extdata
