### 2019-02-05 4baa0e3
Update README.md
added link to vignette as well as direct c++ code in usage
