+++ README.md
@@ -11,6 +11,15 @@ This package provides R with access to [SeqAn v2.4](http://www.seqan.de/) header
 
 RSeqAn can be used via the `LinkingTo:` field in the `DESCRIPTION` field of an R package. Note that the compiler will need to compile as C++14. This can be done in R as `Sys.setenv("PKG_CXXFLAGS"="-std=c++14")`.
 
+For a quick example, see our [introductory vignette](https://compbiocore.github.io/RSeqAn/first_example/). Alternatively you can run the C++ code from the vignette directly by entering
+
+```r
+Rcpp::sourceCpp(system.file("extdata","first_example.cpp",package="RSeqAn"))
+pattern_search("This is an awesome tutorial to get to know SeqAn!", "tutorial")
+```
+
+Don't forget to set the C++14 compiler flags if you do this!
+
 ## Installation
 
 We recommend installing the release version [through Bioconductor](https://bioconductor.org/packages/release/bioc/html/RSeqAn.html). Instructions are provided on their page.
@@ -19,7 +28,7 @@ If you have `devtools`, then the development version can be installed from here 
 ```r
 devtools::install_github("compbiocore/RSeqAn")
 ```
-This version contains `Rcpp::as` and `Rcpp::wrap` functions for `CharString`.
+The development version contains `Rcpp::as` and `Rcpp::wrap` functions for `CharString`.
 
 ## Author
 
@@ -32,6 +41,6 @@ See [`NEWS`](/inst/NEWS) for release history.
 ## Citation
 
 To cite this package from within R, enter `citation("RSeqAn")`. Citation is also provided below:
-	Guang A (2018). *RSeqAn: R SeqAn.* R package version 1.2.0.
+	Guang A (2018). *RSeqAn: R SeqAn.* R package version 1.3.1.
 
 Please do not forget to cite SeqAn if you use this package in your acedemic work, even if only in auxiliary tools or pipelines: https://doi.org/10.1016/j.jbiotec.2017.07.017
