### 2019-02-05 #1160 @aguang
@pdiakumis I'm glad you found RSeqAn useful! Thanks for your comments, I've added a note under Usage with a link to the vignette as well as the C++ code from it as its own example file that can be loaded. I also updated the Installation in the README and created a tagged release for 1.3.1 which hopefully should clear up what version the JOSS paper is referring to. And it looks like the pdf is compiled now.
