### 2019-02-05 #1160 @brainstorm
@aguang Can you please fix the above issues with the DOIs?
