### 2019-02-06 0e104be
updated DOIs
