+++ paper/paper.bib
@@ -25,28 +25,36 @@ title = {{Extending R with C++: A Brief Introduction to Rcpp}},
 year = {2018}
 }
 
-@Article{Huber2015,
-    author = {W. Huber and V. J. Carey and R. Gentleman and S. Anders and M. Carlson and B. S. Carvalho and H. C. Bravo and S. Davis and L. Gatto and T. Girke and R. Gottardo and F. Hahne and K. D. Hansen and R. A. Irizarry and M. Lawrence and M. I. Love and J. MacDonald and V. Obenchain and A. K. {Ole's} and H. {Pag`es} and A. Reyes and P. Shannon and G. K. Smyth and D. Tenenbaum and L. Waldron and M. Morgan},
-    title = {{O}rchestrating high-throughput genomic analysis with {B}ioconductor},
-    journal = {Nature Methods},
-    year = {2015},
-    volume = {12},
-    number = {2},
-    pages = {115--121},
-    url = {http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html},
-  }
-  
-@article{Langmead2013,
-abstract = {As the rate of sequencing increases, greater throughput is demanded from read aligners. The full-text minute index is often used to make alignment very fast and memory-efficient, but the approach is ill-suited to finding longer, gapped alignments. Bowtie 2 combines the strengths of the full-text minute index with the flexibility and speed of hardware-accelerated dynamic programming algorithms to achieve a combination of high speed, sensitivity and accuracy.},
-author = {Langmead},
-doi = {10.1038/nmeth.1923.Fast},
-issn = {1548-7091},
-journal = {Nature methods},
-pmid = {22388286},
-title = {{Bowtie2}},
-year = {2013}
+@article{Huber2015,
+abstract = {Bioconductor is an open-source, open-development software project for the analysis and comprehension of high-throughput data in genomics and molecular biology. The project aims to enable interdisciplinary research, collaboration and rapid development of scientific software. Based on the statistical programming language R, Bioconductor comprises 934 interoperable packages contributed by a large, diverse community of scientists. Packages cover a range of bioinformatic and statistical applications. They undergo formal initial review and continuous automated testing. We present an overview for prospective users and contributors.},
+archivePrefix = {arXiv},
+arxivId = {arXiv:gr-qc/9809069v1},
+author = {Huber, Wolfgang and Carey, Vincent J. and Gentleman, Robert and Anders, Simon and Carlson, Marc and Carvalho, Benilton S. and Bravo, Hector Corrada and Davis, Sean and Gatto, Laurent and Girke, Thomas and Gottardo, Raphael and Hahne, Florian and Hansen, Kasper D. and Irizarry, Rafael A. and Lawrence, Michael and Love, Michael I. and MaCdonald, James and Obenchain, Valerie and Oles̈, Andrzej K. and Pag{\`{e}}s, Herv{\'{e}} and Reyes, Alejandro and Shannon, Paul and Smyth, Gordon K. and Tenenbaum, Dan and Waldron, Levi and Morgan, Martin},
+doi = {10.1038/nmeth.3252},
+eprint = {9809069v1},
+isbn = {1548-7091},
+issn = {15487105},
+journal = {Nature Methods},
+pmid = {25633503},
+primaryClass = {arXiv:gr-qc},
+title = {{Orchestrating high-throughput genomic analysis with Bioconductor}},
+year = {2015}
 }
 
+@article{Langmead2012,
[…]
