### 2019-02-07 #1160 @aguang
actually there appears to already have been a Zenodo doi made for me when I made a release tag on Github: 10.5281/zenodo.2558617. Not sure which one would be better to use.
