### 2019-02-07 #1160 @kyleniemeyer
Hello @aguang, before we accept and publish your submission, could you please make some small edits to the paper? In particular, the article could use some revision at the very beginning, to satisfy these requirements:
- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software

(What you have now isn't too far off, but I admit as someone whose field is far from that relevant to the software, I don't quite understand what this software—or SeqAn, really—is used for. Some general description or examples of use cases in that paragraph would also help.)

Thanks!
