+++ paper/paper.md
@@ -28,7 +28,7 @@ R is used heavily in the bioinformatics community for processing, analyzing, and
 
 Writing certain functions in C++ also allows the researcher to take advantage of high-performance C++ libraries. One such library that has proven powerful and popular for bioinformatics applications is SeqAn [@Doring2008; @Reinert2017]. SeqAn is an open source C++ library of efficient algorithms and data structures for the analysis of sequences with a focus on biological data. It has been used for many popular bioinformatics tools, including Bowtie2 [@Langmead2012] and Tophat [@Trapnell2009]. Its capabilities include efficient storage and I/O of sequence data, fast algorithms for pattern matching, and much more.
 
-To date, no R package has taken advantage of SeqAn. This is due to two issues. First, R for Windows is built with the mingw compiler, which the SeqAn development team does not offer support for. Second, SeqAn implements several custom data types in order to make the library more efficient. This means that any data type in SeqAn needs to have have R to C++ conversion and C++ to R conversion functions written for it before it can be used in R.
+To date, no R package has taken advantage of SeqAn. This is due to two issues. First, R for Windows is built with the mingw compiler, which the SeqAn development team does not offer support for. Second, SeqAn implements several custom data types in order to make the library more efficient. This means that any data type in SeqAn needs to have R to C++ conversion and C++ to R conversion functions written for it before it can be used in R.
 
 `RSeqAn` solves both of these problems, thus allowing any R researcher to use the capabilities of SeqAn in their own work without need for the user to install SeqAn themselves or deal with interfacing between R and C++. `RSeqAn` can be installed from Bioconductor, and documentation for it can be found in the package as well as [online](https://compbiocore.github.io/RSeqAn/).
 
