### 2019-02-11 #1160 @brainstorm
@arfon, I think we are done and ready to accept here :)
