### 2019-02-12 #1160 @arfon
@brainstorm - please ping @openjournals/joss-eics in the future when handing off the submissions. We now have one of us on a rotation.
